Replace 'text' to 'content' module key
Using 'content' as key for the tag content instead of text will allow
to change the content of the model in future with a non only text.

Such as html or numbers or something else.

Also, this is more generic and can be standardized in future when all of
the data entries will become Data objects.